+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 - The "unload" window event has been replaced with "pagehide" (due to "unload" being deprecated).
 - The public function "setConfiguration()" will now only update the configuration if something in the configuration has been changed.
 - All public functions now only accept the right variable types for data (to prevent issues from occurring).
+- Added "use strict" support internally and updated all public functions to use the new scope.
 
 <br>